Add Classification directory Move Classification store code into its own directory. Move ClassificationMetadata into its own file.
