+++ packages/lib-classifier/src/store/Classification/index.js
@@ -1,0 +1,2 @@
+export { default } from './Classification'
+export { default as ClassificationMetadata } from './ClassificationMetadata'
